requirements.txt
aiogram==2.25.1
tortoise-orm==0.20.0
asyncpg==0.29.0
python-dotenv==1.0.1